converted vm type to interface
