+++ codegen/codegen.go
@@ -11,13 +11,7 @@ import (
 	"time"
 
 	"github.com/HexmosTech/gabs/v2"
-	"github.com/HexmosTech/lama2/cmdexec"
-	"github.com/dop251/goja"
-	"github.com/rs/zerolog/log"
 )
-
-//go:embed httpsnippet.js
-var snippetcore string
 
 type SnippetArgs struct {
 	Language    string
@@ -25,14 +19,6 @@ type SnippetArgs struct {
 	HARRequest  string
 	SnippetCore string
 }
-
-var globalVM *goja.Runtime
-
-func initialize() {
-	globalVM = cmdexec.GetJSVm()
-}
-
-var flag = 0
 
 func GetRequestHARString(block *gabs.Container) (string, int) {
 	httpv := block.S("verb", "value")
@@ -160,45 +146,4 @@ func PrepareHTTPSnippetGenerator(snippetArgs SnippetArgs) string {
 	tmpl, _ := template.New("httpsnippet").Parse(templStr)
 	tmpl.Execute(&templOutput, snippetArgs)
 	return templOutput.String()
-}
-
-func generateConvertedSippet(targetLangLib string, parsedAPI *gabs.Container) string {
-	initialize()
-	parsedAPIblocks := parsedAPI.S("value").Data().(*gabs.Container).Children()
-	convertedSnippetList := make([]string, 0)
-
-	for i, block := range parsedAPIblocks {
-		log.Debug().Int("Block num", i).Msg("")
-		log.Debug().Str("Block getting processed", block.String()).Msg("")
-		blockType := block.S("type").Data().(string)
-		if blockType == "processor" {
-			snippet := block.S("value").Data().(*gabs.Container).Data().(string)
-			log.Debug().Str("Processor block incoming block", block.String()).Msg("")
-			convertedSnippetList = append(convertedSnippetList, snippet)
-		} else if blockType == "Lama2File" {
-			harRequest, flag := GetRequestHARString(block)
-			snippetArgs := SnippetArgs{}
-			lang, lib := SplitLangLib(targetLangLib)
-			snippetArgs.Language = lang
-			snippetArgs.Library = lib
-			snippetArgs.HARRequest = harRequest
-			snippetArgs.SnippetCore = snippetcore
-			httpsnippetCode := PrepareHTTPSnippetGenerator(snippetArgs)
-			vm := cmdexec.GetJSVm()
-			_, e := vm.RunString(httpsnippetCode)
-			if e != nil {
-				log.Fatal().
-					Str("Type", "CodeGen").
-					Str("Error", e.Error()).
-					Msg("Code generator error")
-			}
-			convertedSnippet := vm.Get("convertedSnippet").String()
-			convertedSnippetWithPostProcessing := postprocessURL(convertedSnippet, flag)
-			flag = 0
-			convertedSnippetList = append(convertedSnippetList, convertedSnippetWithPostProcessing)
-		}
-	}
-
-	convertedSnippetFinal := strings.Join(convertedSnippetList, "\n")
-	return convertedSnippetFinal
 }
